+++ src/chainer.rs
@@ -11,11 +11,11 @@ use crate::nam::Nam;
 
 const N_PRECOMPUTED: usize = 1024;
 
-#[derive(Debug, Ord, PartialOrd, Eq, PartialEq)]
-struct Anchor {
-    ref_id: usize,
-    ref_start: usize,
-    query_start: usize,
+#[derive(Debug, Ord, PartialOrd, Eq, PartialEq, Clone, Copy)]
+pub struct Anchor {
+    pub ref_id: usize,
+    pub ref_start: usize,
+    pub query_start: usize,
 }
 
 #[derive(Debug)]
@@ -198,6 +198,21 @@ impl Chainer {
             trace!("Chaining {} anchors", anchors.len());
             anchors.sort_unstable_by_key(|a| (a.ref_id, a.ref_start, a.query_start));
             anchors.dedup();
+
+            // trace!(
+            //     "Anchors for {} strand [{}]",
+            //     if is_revcomp == 0 {
+            //         "forward"
+            //     } else {
+            //         "reverse"
+            //     },
+            //     anchors
+            //         .iter()
+            //         .map(|anchor| format!("{:?}", anchor))
+            //         .collect::<Vec<_>>()
+            //         .join(",")
+            // );
+
             let (best_score, dp, predecessors) = self.collinear_chaining(&anchors);
 
             extract_chains_from_dp(
@@ -378,6 +393,7 @@ fn extract_chains_from_dp(
         let mut j = i;
         let mut c = 1;
         let mut overlaps = false;
+        let mut chain_anchors = vec![anchors[i]];
 
         let mut matching_bases = k;
         let mut ref_coverage = anchors[i].ref_start;
@@ -388,6 +404,7 @@ fn extract_chains_from_dp(
                 overlaps = true;
                 break;
             }
+            chain_anchors.push(anchors[j]);
             used[j] = true;
             c += 1;
 
@@ -401,6 +418,7 @@ fn extract_chains_from_dp(
 
         let first = &anchors[j];
         let last = &anchors[i];
+
         chains.push(Nam {
             nam_id: chains.len(),
             query_start: first.query_start,
@@ -412,6 +430,7 @@ fn extract_chains_from_dp(
             ref_id: last.ref_id,
             score: score + c as f32 * chaining_parameters.matches_weight,
             is_revcomp,
+            anchors: chain_anchors,
         });
     }
 }