+++ src/mapper.rs
@@ -19,6 +19,7 @@ use crate::insertsize::InsertSizeDistribution;
 use crate::math::normal_pdf;
 use crate::mcsstrategy::McsStrategy;
 use crate::nam::{Nam, get_nams_by_chaining, reverse_nam_if_needed};
+use crate::piecewisealigner::remove_spurious_anchors;
 use crate::read::Read;
 use crate::revcomp::reverse_complement;
 use crate::sam::{
@@ -37,6 +38,7 @@ pub struct MappingParameters {
     pub max_tries: usize,
     pub mcs_strategy: McsStrategy,
     pub output_unmapped: bool,
+    pub use_ssw: bool,
 }
 
 impl Default for MappingParameters {
@@ -48,6 +50,7 @@ impl Default for MappingParameters {
             max_tries: 20,
             mcs_strategy: McsStrategy::default(),
             output_unmapped: true,
+            use_ssw: false,
         }
     }
 }
@@ -395,7 +398,7 @@ pub fn align_single_end_read(
     aligner: &Aligner,
     rng: &mut Rng,
 ) -> (Vec<SamRecord>, Details) {
-    let (nam_details, nams) = get_nams_by_chaining(
+    let (nam_details, mut nams) = get_nams_by_chaining(
         &record.sequence,
         index,
         chainer,
@@ -424,7 +427,11 @@ pub fn align_single_end_read(
 
     let k = index.parameters.syncmer.k;
     let read = Read::new(&record.sequence);
-    for (tries, nam) in nams.iter().take(mapping_parameters.max_tries).enumerate() {
+    for (tries, nam) in nams
+        .iter_mut()
+        .take(mapping_parameters.max_tries)
+        .enumerate()
+    {
         let score_dropoff = nam.n_matches as f32 / nam_max.n_matches as f32;
 
         if (tries > 1 && best_edit_distance == 0)
@@ -437,9 +444,46 @@ pub fn align_single_end_read(
             details.inconsistent_nams += 1;
             continue;
         }
-        let Some(alignment) = extend_seed(aligner, nam, references, &read, consistent_nam) else {
+        let Some(alignment) = extend_seed(
+            aligner,
+            nam,
+            references,
+            &read,
+            consistent_nam,
+            mapping_parameters.use_ssw,
+        ) else {
             continue;
         };
+
+        // outputting Piecewise vs SSW alignments for debugging
+        // if log::log_enabled!(log::Level::Trace) {
+        //     let (mut ssw, mut pw) = if !mapping_parameters.use_ssw {
+        //         (
+        //             extend_seed(aligner, nam, references, &read, consistent_nam, true).unwrap(),
+        //             alignment.clone(),
+        //         )
+        //     } else {
+        //         (
+        //             alignment.clone(),
+        //             extend_seed(aligner, nam, references, &read, consistent_nam, false).unwrap(),
+        //         )
+        //     };
+        //     // manually adds the soft clips
+        //     let mut cigar = Cigar::new();
+        //     cigar.push(CigarOperation::Softclip, pw.soft_clip_left);
+        //     cigar.extend(&pw.cigar);
+        //     cigar.push(CigarOperation::Softclip, pw.soft_clip_right);
+        //     pw.cigar = cigar;
+        //
+        //     let mut cigar = Cigar::new();
+        //     cigar.push(CigarOperation::Softclip, ssw.soft_clip_left);
+        //     cigar.extend(&ssw.cigar);
+        //     cigar.push(CigarOperation::Softclip, ssw.soft_clip_right);
+        //     ssw.cigar = cigar;
+        //
+        //     trace!("Alignment:[{:?},SSW:{:?},PW:{:?}]", nam.clone(), ssw, pw);
+        // }
+
         details.tried_alignment += 1;
         details.gapped += alignment.gapped as usize;
 
@@ -530,10 +574,11 @@ pub fn align_single_end_read(
 /// alignment.
 fn extend_seed(
     aligner: &Aligner,
-    nam: &Nam,
+    nam: &mut Nam,
     references: &[RefSequence],
     read: &Read,
     consistent_nam: bool,
+    use_ssw: bool,
 ) -> Option<Alignment> {
     let query = if nam.is_revcomp {
         read.rc()
@@ -576,11 +621,18 @@ fn extend_seed(
         }
     }
     if gapped {
-        let ref_start = projected_ref_start.saturating_sub(50);
-        let ref_end = min(projected_ref_end + 50, refseq.len());
-        let segment = &refseq[ref_start..ref_end];
-        info = aligner.align(query, segment)?;
-        result_ref_start = ref_start + info.ref_start;
+        let padding = read.len() / 10;
+        if use_ssw {
+            let ref_start = projected_ref_start.saturating_sub(padding);
+            let ref_end = min(projected_ref_end + padding, refseq.len());
+            let segment = &refseq[ref_start..ref_end];
+            info = aligner.align(query, segment)?;
+            result_ref_start = ref_start + info.ref_start;
+        } else {
+            remove_spurious_anchors(&mut nam.anchors);
+            info = aligner.align_piecewise(query, refseq, &nam.anchors, padding)?;
+            result_ref_start = info.ref_start;
+        }
     }
     Some(Alignment {
         cigar: info.cigar.clone(),
@@ -803,8 +855,22 @@ fn extend_paired_seeds(
         let consistent_nam2 = reverse_nam_if_needed(&mut n_max2, read2, references, k);
         details[1].inconsistent_nams += !consistent_nam2 as usize;
 
-        let alignment1 = extend_seed(aligner, &n_max1, references, read1, consistent_nam1);
-        let alignment2 = extend_seed(aligner, &n_max2, references, read2, consistent_nam2);
+        let alignment1 = extend_seed(
+            aligner,
+            &mut n_max1,
+            references,
+            read1,
+            consistent_nam1,
+            true, // SSW
+        );
+        let alignment2 = extend_seed(
+            aligner,
+            &mut n_max2,
+            references,
+            read2,
+            consistent_nam2,
+            true, // SSW
+        );
         if let (Some(alignment1), Some(alignment2)) = (alignment1, alignment2) {
             details[0].tried_alignment += 1;
             details[0].gapped += alignment1.gapped as usize;
@@ -832,7 +898,14 @@ fn extend_paired_seeds(
     for i in 0..2 {
         let consistent_nam = reverse_nam_if_needed(&mut nams[i][0], reads[i], references, k);
         details[i].inconsistent_nams += !consistent_nam as usize;
-        a_indv_max[i] = extend_seed(aligner, &nams[i][0], references, reads[i], consistent_nam);
+        a_indv_max[i] = extend_seed(
+            aligner,
+            &mut nams[i][0],
+            references,
+            reads[i],
+            consistent_nam,
+            true, // SSW
+        );
         details[i].tried_alignment += 1;
         details[i].gapped += a_indv_max[i].as_ref().map_or(0, |a| a.gapped as usize);
         alignment_cache[i].insert(nams[i][0].nam_id, a_indv_max[i].clone());
@@ -863,8 +936,14 @@ fn extend_paired_seeds(
                     let consistent_nam =
                         reverse_nam_if_needed(&mut this_nam, reads[i], references, k);
                     details[i].inconsistent_nams += !consistent_nam as usize;
-                    alignment =
-                        extend_seed(aligner, &this_nam, references, reads[i], consistent_nam);
+                    alignment = extend_seed(
+                        aligner,
+                        &mut this_nam,
+                        references,
+                        reads[i],
+                        consistent_nam,
+                        true, // SSW
+                    );
                     details[i].tried_alignment += 1;
                     details[i].gapped += alignment.as_ref().map_or(0, |a| a.gapped as usize);
                     e.insert(alignment.clone());
@@ -964,7 +1043,8 @@ fn rescue_read(
         }
         let consistent_nam = reverse_nam_if_needed(nam, read1, references, k);
         details[0].inconsistent_nams += !consistent_nam as usize;
-        if let Some(alignment) = extend_seed(aligner, nam, references, read1, consistent_nam) {
+        if let Some(alignment) = extend_seed(aligner, nam, references, read1, consistent_nam, true)
+        {
             details[0].gapped += alignment.gapped as usize;
             alignments1.push(alignment);
             details[0].tried_alignment += 1;